- 회의 참가, 종료 시각, 회의 시간 기록 기능 추가

+++ src/main/java/com/mindplates/everyonesprint/biz/meeting/entity/Meeting.java
@@ -48,7 +48,7 @@ public class Meeting extends CommonEntity {
     private LocalDateTime realEndDate;
 
     @Column(name = "duration_seconds")
-    private Integer durationSeconds;
+    private Long durationSeconds;
 
     @OneToMany(fetch = FetchType.EAGER, mappedBy = "meeting", cascade = CascadeType.ALL, orphanRemoval = true)
     @Fetch(value = FetchMode.SELECT)
