+++ src/main/java/com/mindplates/everyonesprint/biz/meeting/vo/response/MeetingListResponse.java
@@ -15,7 +15,7 @@ public class MeetingListResponse {
     private LocalDateTime startDate;
     private LocalDateTime endDate;
     private Integer userCount;
-    private Integer durationSeconds;
+    private Long durationSeconds;
 
     public MeetingListResponse(Meeting meeting) {
         this.id = meeting.getId();
